Motion_detectioN_using_opencv/requirements.txt
bokeh==3.6.1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
tornado==6.4.2
tzdata==2024.2
urllib3==2.2.3
xyzservices==2024.9.0
